Rebase the initial implementation of Page Graph
It looks like the Page Graph commits before this point were improperly
rebased at some point, and their patches no longer apply to the version
of the Chromium source tree they're tied to. Now that the original
context is lost, we can't really salvage that part of the original
commit history. So I've smooshed those commits together into this one
which *does* apply properly, and which marks the new starting point of
the Page Graph branch.