Do not immediately expire sync v1 infobar about v2
We were relying on the base implementation. Possibly i don't understand how that works because I had assumed that it would only expire the infobar after a change of site, not a New Tab or resurrecting a tab on browser restart. But in release (and frustratingly not on local builds) that's not the case - the infobar was immediately dismissing. So, change the logic to an info bar which does not automatically expire for the active tab.