Removed unneeded patch.
This patch was an upstream patch from Chromium 85 that was needed to
build Chromium 84 with Windows 10.0.19041 SDK.

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/ab3b65fecd8ddfd1c857567b2f358addae6679bf

commit ab3b65fecd8ddfd1c857567b2f358addae6679bf
Author: Bruce Dawson brucedawson@chromium.org
Date: Fri Jun 5 20:43:15 2020 +0000

Make the Windows 10.0.19041 SDK compatible with clang-cl

The Windows 10.0.19041 SDK modifies module.h so that it passes a lambda
to InitOnceExecuteOnce. This works poorly because it is not possible to
specify the calling convention of a lambda, and the callback requires a
specific and non-default calling convention.

This code works with VC++ because VC++ lambdas provide all possible
calling conventions. It works with 64-bit builds with clang-cl because
there is only one calling convention. But it fails on 32-bit builds with
clang-cl.

This define tells module.h to use the old singleton technique. The
Microsoft developer who wrote the new technique has been made aware. See
the bug for details.

This change is a NOP for packaged toolchain users until we switch the
packaged toolchain to the new SDK, but it avoids problems for developers
who install the new SDK and build with it.

Bug: 1089996