Fixed default SR topsites are disappeared when toggling themes option
When option is changed from SR to default, SR's topsites was deleted.
This PR prevent that deleting by introducing defaultSRTopSite property
into each item. If it's set to true, we will not delete those items
even if history doesn't have it. It will be only deleted when user
explicitely deletes.