Code Review: turn off PGO.
Don't download profiles either.

PGO was enabled for official builds for Mac and Windows in Cr84, but the
profiles weren't being downloaded. Downloading profiles was added to
tools in the Cr85 timeframe. For now we'll turn off PGO and will need a
followup to determine if we should turn it back on.