Eliminate race conditions from FinalPageGraph
Route the IPC up through to the RenderProcessHost directly, in case the
RenderView loses connection to the RenderViewHost before the message is
sent. Have RenderProcessHost find the correct main frame based on the
devtools main frame token and dispatch Page.finalPageGraph events
directly to the associated devtools agents.

Register documents with Page Graphs as having an unload handler, to
try and prevent the render process from being killed before the
FinalPageGraph IPC message is completely dispatched.