[Android] MenuButton is hidden behind MenuButtonCoordinator (additional changes)
Chromium change:
https://chromium.googlesource.com/chromium/src/+/b779b577c3dd156f9bfda3f6ca72fc2c58e29441

Reland:[ToolbarMVC] Hide MenuButton interactions in toolbar.top behind MenuButtonCoordinator

Most method calls are translated 1:1 by forwarding the call to the equivalent
MenuButton or View method. A notable exception is handling of the tab switcher
animation, which currently requires a view in order to perform canvas
translation before drawing the button in its animated state. In this case,
we provide a direct accessor for the view.

Reland note: the NPE was due to the MenuButtonCoordinator not being set until
right after inflation finishes, which is when ToolbarTablet was trying to use
it.

Bug: 1086676