Use full relative path to Brave module files from generated V8 bindings
The new IDL compiler removes the ../../ prefix for include files in an
attempt to "canonicalize the paths heuristically", which works well for
modules under //third_party/blink/renderer/modules.

Unfortunately, that does not work well with our Brave module since it
lives under //brave/third_party/blink/renderer/modules, and the prefix
needs to be kept so that header files can be correctly included from
the generated sources' directory.

This patch adapts the python code that does this adjustment so that
it does not remove the prefix for Brave-specific files.