don't detect window.ethereum usage when enumerated
certain sites like wayback machine will enumerate named properties of
`window`, causing a false positive detection on `window.ethereum`,
showing an info bar to install the Crypto Wallets extension for sites
that don't actually require it.

this fix is a bit broder scoping and rough in nature, as it relies on
checking the code of the surrounding call-site to see if
`getOwnPropertyNames` is used. there are many possible ways that
`getOwnPropertyNames` could be seen in source code that don't result in
enumeration of `window`, but it is best to be more conservative for this
automatic detection behavior

see brave/brave-browser#11865 for more discussion

Fixes brave/brave-browser#11865