[Android] MenuButton related changes
Chromium changes:
https://chromium.googlesource.com/chromium/src/+/a65f80f87cc151be897131164e015d37bea1d155

[ToolbarMVC] Remove direct MenuButton interactions from Tabswitcher

This only requires giving TabSwitcherModeTTCoordinatorPhone a
MenuButtonCoordinator and setting the MenuButton view once inflation happens.
The logic removed (tinting, setting the onClick/accessibilityDelegate)
is now handled internally by MenuButtonCoordinator/MenuButton.

Bug: 1086676

https://chromium.googlesource.com/chromium/src/+/fbc145eeea76493c5469953975322db185826e3f

[ToolbarMVC] Reimplement MenuButton internals to follow MVC

This adds the familiar MVC pieces, in addition to the existing Coordinator and View: a Mediator, a ViewBinder, and a Properties class. Logic is moved out of the MenuButton and MenuButtonCoordinator to the Mediator. Of note, there are two compound properties for combinations of properties that need to be set together:
* show/hide badge and shouldAnimate
* useLight and ColorStateList

Bug: 1086676