Greaselion accepts "minimium_brave_version" config property and "supports-minimum-brave-version" precondition
Allows scripts to specify which minimum version of brave they are compatible with (supports wildcards). Also contains another precondition so that browsers which do not support this new config property can ignore scripts targeted to specific minimum brave versions. That new precondition can be temporary and removed in a few months when everyone is upgraded.