Greaselion validates rule validity
Previously, the added browser test case (where scripts or urls are not defined) would cause a difficult-to-debug browser crash due to not checking before passing nullptr as ListValue* to Parse. This lets those rules be parsed, but adds an IsValid method to check for minimum requirements we need to run a rule.