New Tab Page: Allow page to scroll before Today kicks in and fix backdrop-filter
Also, tighten style for indicator.
Overcomes a bug where a backdrop-filter element's ancestor which has a filter must also have a background. When this bug is fixed then this element won't need the background. Before this commit, the Widget title backgrounds would not correctly have backdrop filter because their ancestor had a blur and not also a background (it came from a higher up ancestor).