[Android] Return type of Fragment.getActivity() changed to Activity
Chromium change:
https://chromium.googlesource.com/chromium/src/+/ae52b2a4a4f3fdf121619fc62f300883b610bd46

[WebLayer] Change the return type of Fragment.getActivity() to Activity.

This CL adds the ability to specify a script in an android_aar_prebuilt
rule that can modify prebuilt jar files, and adds a script to change
Fragment.getActivity() and Fragment.requireActivity() to return an
Activity instead of a FragmentActivity.

This is the first CL in a chain that will allow us to remove the fake
activity we create when embedding Fragments that cross classloader
boundaries.

Bug: 1123216