7955: Adapt PCSTor::CircuitIsolationKey to using net::NetworkIsolationKey
This patch changes ProxyConfigServiceTor::CircuitIsolationKey() to rely
on upstream's net::NetworkIsolationKey() instead of implementing pretty
much the same logic manually.

Note that net::NetworkIsolationKey() does not provide a way to retrieve
its internal |top_frame_origin_|, which stores the origin/etld+1 of the
original top frame passed on constructions, which is what we need for
the CircuitIsolationKey according to the Tor design specs. For this
reason, this patch also provide a chromium_src override for the file
//net/base/network_isolation_key.h that adds an extra helper for that.