Make sure cookies-related changes are notified for Brave Shields
BravePrefProvider's UpdateCookieRules() expects a ContentSettingsType
parameter to determine whether BravePrefProvider::NotifyChanges() should
be called, which used to be either ContentSettingsType::COOKIES or
ContentSettingsType::PLUGINS before the removal of ResourceIdentifier.

In the case of that parameter being ContentSettingsType::PLUGINS, the
UpdateCookieRules() method would only be called when notifying of a
change related to Brave Shields itself (i.e. shields up or down) or
to a change especific to Brave Shields' cookies, in which case the
UpdateCookieRules() would notify as ContentSettingsType::COOKIES.

However, now that ResourceIdentifier is gone and we're using Brave
specific ContentSettingTypes, we need to update UpdateCookieRules()
to make sure that we report as ContentSettingsType::COOKIES when
either a content type of ContentSettingsType::BRAVE_COOKIES or
ContentSettingsType::BRAVE_SHIELDS gets passed, so that we maintain
the same functionality as before.