Don't cleanup "plugins" preference before migrating Shields' settings
Calling MigrateShieldsSettings from BravePrefProvider's constructor is
too late if we want to make sure we have a chance to migrate settings
that have been deprecated upstream, before they are wiped out from the
PrefStore, in PrefProvider::DiscardOrMigrateObsoletePreferences().

Since that method is called from PrefProvider's constructor, we can't
simply override it and from BravePrefProvider as such override would
not be called from the base class' constructor, so we add a small
change to defer clearing up that particular setting until we have
migrated Shields settings instead.