Don't add ContentSettingsTypes for FB, Twitter and LinkedIn embeds
At the beginning of this migration I converted everthing listed inside
brave_shield_constants.h into a Brave-specific ContentSettingsType,
but after working on this migration I've realized it does not really
make sense, since we're not storing shields-specific settings for them,
and so they can be dropped.

We need to keep the kFacebookEmbeds, kTwitterEmbeds and kLinkedInEmbeds
though, as those are still required by the ad block service.