[Android] Fixed assert for non-existing preference
profile.content_settings.exceptions.plugins is no longer a registered
path for Android (since [1]), but we still need to read it to migrate
Shields settings, so let's register it manually ourselves.

Additionally, this CL moves the cleanup of the migration-related paths
to MigrateShieldsSettingsFromResourceIds(), which seems a better place
than doing it from MigrateShieldsSettings(), which is not only related
to this particular migration away from ResourceIdentifier.

[1] https://crrev.com/c//2577782