P3A: Introduce P3ACollector so that we may retrieve accurate values in time for metric sends
Allows a service to register itself as an "observer" and ensure that accurate metrics are provided before P3A begins to stage and transmit the last "period" (week) of data. This overcomes the issue whereby no user action has happened for a metric within a time period, and the metric never being reset to "0", for example.