Bring CloseGuestProfileWindows() back into brave_profile_menu_view.cc
This method has been removed from upstream as it was not used and there
does not seem to be an equivalent one available, so let's bring it back.

Fixup! Mario says: on a second thought I think this is probably not
needed so reverting this change (and simply dropping the
CloseGuestProfileWindows() call from
BraveProfileMenuView::OnExitProfileButtonClicked().

Chromium change:
https://chromium.googlesource.com/chromium/src.git/+/53d0699ee2609b7195a93d92d2433500a0affd55

commit 53d0699ee2609b7195a93d92d2433500a0affd55
Author: Monica Basta <msalama@chromium.org>
Date:   Wed Feb 10 11:57:07 2021 +0000

    [Profiles]: Avoid opening the profile picker on close 'x' windows

    Currently on click on 'Close x windows' from the profile menu, the
    profile is closed and the user manager/profile picker is opened. This
    behavior is no longer desired. Clicking on the button should only close
    the profile. This CL also removes 'CloseGuestProfileWindows' function
    that is not used.

    Bug: 1176599