Drop patch for //c/install_static/install_util.cc and use chromium_src
This patch was only enabling a section of the code via OFFICIAL_BUILD
and adapting the expected channel name for Channel::CANARY to "nightly"
instead, being both changes in the GetChromeChannel() helper function.

Instead of requiring a patch for that, we can simply provide our own
implementation of GetChromeChannel() overriding upstream's, which is
fairly stable and hasn't changed for ages.

Fixes https://github.com/brave/brave-browser/issues/14334