Remove extensions-common-manifest_url_handlers.cc.patch
This can be removed in favour of a chromium_src override to redefine
ManifestURL::GetWebStoreURL() into a method that unconditionally
returns an empty GURL.

Note that we need to also redefine GetHomepageURL() as it contains
a call to GetWebStoreURL(), otherwise upstream's version of the
GetWebStoreURL() function would be still be invoked via that way.
This is not ideal, but at least the code in GetHomepageURL() that
needs to be replicated in the chromium_src is very small and very
stable (hasn't changed in 8 years), so risk should be minimal.

Resolves https://github.com/brave/brave-browser/issues/14343