Fix windows build due to invalid assignment to a constant
While landing PR #8063 [1], we didn't realize that the assignment in
the chromium_src override was wrong because the CI bots didn't complain,
since the installer is not supported in non official builds, and
because I normally build on Linux only.

Then I filed issue 14365 [2] to cleanup all those non-official code
paths from changes to //chrome/installer/** files and, as soon as I
started building on my Windows machine with official_build = false
I hit a build error because of this assignment.

Therefore, while the long term fix will come as part of fixing that
other issue, it would be great to get non-official Windows building
again.

[1] https://github.com/brave/brave-core/pull/8063/files
[2] https://github.com/brave/brave-browser/issues/14364

Resolves https://github.com/brave/brave-browser/issues/14365