Make EXCLUDES regexp list work on non-UNIX-like shells (e.g. Windows)
To make this work well on shells like Windows' cmd.exe or PowerShell,
we add an intermediate step before matching the paths for files found
in the repository against the regular expresisons declared in EXCLUDES,
so that we use forward slashes instead of backward slashes, which will
cause errors due to the regexp interpreting paths like C:\\mypath as if
they contained the escaped '\m' character.

Once the regular expression has been used though, we still append the
unnormalized OS-dependant relative path to the list of paths built by
filter_chromium_src_filepaths(), so that further operations over those
files can be performed correctly, and so that any path printed in the
output will use the OS-dependant path separator.