speedreader: Implement alternative candidates
This is another step towards algorithmic parity with Readability.js.

In the current implementation, we root the new DOM at top scoring
candidate. With this patch, we keep 1 top candidate and 4 alternative
candidates. If 3/4 of those candidates have a score near the top
candidate, we start iterating backwards starting from the top candidate.
If there is a shared parent with any 3 of the alternative candidates, we
make that the new root.

This is important, because some articles on Vanity Fair score very high
with a deeply nested top candidate, and real content is deleted from the
new DOM.