Fix build due to no longer existing render_frame_id in network service
The network service does no longer return a render_frame_id, so we need
to adapt Brave-specific code that was using that piece of information
away from it.

Fortunately, in every single place where this needs to be migrated
Brave seems to use the render_frame_id, along with the ID of the
relevant render process, as a fallback mechanism to obtain a
WebContents object associated to such frame in case that object
could not previously be obtained with the frame_tree_node_id.

However, this fallback mechanism should not be needed because a
FrameTreeNode will already have a pointer to the current RFH that
represents a given render frame in a given process, and from that
RFH object obtaining a WebContents object should always work unless
the RFH was already shutting down, and in that case the fallback
mechanism wouldn't be helpful either.

Therefore, this patch gets rid of all the places where we were using
render_process_id and render_frame_id as a fallback mechanism to
obtain a WebContents when frame_tree_node_id can be used, removing
the dependency.

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/cc266d77d5b5de6f928f43125488d10cc01bee9b

commit cc266d77d5b5de6f928f43125488d10cc01bee9b
Author: Dave Tapuska <dtapuska@chromium.org>
Date:   Tue Mar 16 15:41:29 2021 +0000

    Remove renderer_frame_id from network service.

    All reads of the render_frame_id_ field have been removed so it is now
    no longer necessary to set the value. Remove the field from the
    structures.

    BUG=1173710