speedreader: Remove tree data structures
We are maintaining a very minimal fork of kuchiki-rs at
https://github.com/keur/kuchiki/tree/speedreader. We expose two new
fields to API users: `score` and `is_candidate`. This removes the need
for both the nodes binary tree and the candidates binary tree. Kuchiki
has nice helper methods for accessing parents, descendants, and
ancestors, further removing the need for the Paths. The trees indexed by
`Path` structs were mostly hacking around the lack of API extensibility.