speedreader: Unwrap <div> with single <p> node
Some sites, like Slate and the BBC, have <div> elements enclosing single
<p> nodes. This weakens the paragraph's score from propagating higher in
the DOM during the scoring step. In the pre-process step, make this
replacement.

TODO: We may also want to consider doing the same thing for divs with no
block-level elements. I didn't implement it here because the function
_hasChildBlockElement() in Readability.js checks for non-block elements,
so the function in confusing and may be buggy? Not sure.