speedreader: Replace e2e tests with slideshow
The problem with the current tests is that any minor change to the
alogrithm requires a large diff of the newly generated HTML. Since we
test individual parts of the algorithm in doc tests, we make the E2E
tests an HTML slideshow. Anyone making changes to the algorithm can
reference the before and after renders of many sites with low friction.

Resolves https://github.com/brave/brave-browser/issues/15102