speedreader: Start handling lazy-loaded images
Lazy loaded images leave out the src and srcset attributes, and set
other metadata attributes that are fetched and set in JavaScript.
Implement some simple heuristics to do the this in the clean pass and,
if the image can't be loaded, discard it.

Partially resolves https://github.com/brave/brave-browser/issues/15101