Fixed crash from typing in tor window when SE extension is active
fix https://github.com/brave/brave-browser/issues/15140

When OnTemplateURLServiceChanged() is called, it tries to access default
search engine provider but it's null when search engine extension is installed
and that extension enabled to private window.
Those code is not valid now so just deleted because we can't change tor window's
search engine during the run time. It was possible ago but not now.