Element Picker initial implementation
This is the first pass of the element picker. Let's discuss how it
operates.

Overview
--------
When the user activates the picker via the context menu, the element
picker iframe is appended to the document.documentElement from the
content script. We apply some CSS rules so the iframe overlays the
entire page and handles all pointer events.  The latter is important so
that links and buttons don't compete with the element picker UI. The
former is important so that points inside the iframe are the same as
points on the parent page.

Drawing element bounding boxes
------------------------------
When the user hovers around the page, bounding boxes are drawn around
"target" elements. Since the iframe consumes all pointer events, it
sends a message to the content script with the coordinates of the cursor.
The element on the source page is fetched with document.elementFromPoint(..).
The candidate element has its bounding box coordinates sent back to the
iframe. The box is drawn by applying a <rect> masks on an otherwise
transparent black background.

Selecting an element
--------------------
When a user clicks a bounding box, once again, the coordinates are used
to fetch an element on the source page. We create a CSS selector from
the element id, classes, and tree hierachy, applying the process
recursively on parent nodes until we reach the body. We use this
selector and fetch all matching elements; the bounding box coordinates
of all matching elements are sent to the iframe. Drawing is done the same
way as discussed in the previous section.

Creating the filter
-------------------
Once a user has selected an element, they will see orange bounding
boxes on the screen, representing all elements that match the filter
rule. Upon selecting "Create", the cosmetic filter is applied and saved
for that site.

Why go through the trouble of an iframe?
----------------------------------------
(1) It helps us isolate click events from the source page.
(2) Popping up the filter selection on the source page would mean that
    the source page could run JavaScript and modify the cosmetic filter rules
    before the user clicked "Create". We set the src of the iframe to be a
    web accessible resource, since a nice property of these is the uri
    is chrome-extension://. The source page will always have a protocol of
    http://, it will never be able to modify the contents of the iframe due to
    the same-origin-policy.

Future TODOs
------------
  * Make the UI nicer (I'm not a gifted UI person and could use some help)
  * Add a "Preview" button to the picker UI that temporarily applies the
    cosmetic filter rule.
  * Implement the uBlock "specificity" algorithm. This runs heuristics
    to make selector less generic. Used for matching elements that seem
    similar to the target.

Resolves https://github.com/brave/brave-browser/issues/13808