Adapt the way we extend ProfileManager's GetLastUsedProfileName()
Adapt the way we extend ProfileManager's GetLastUsedProfileName()

This used to be a method of ProfileManager we could override from our
BraveProfileManager class, but now it has been moved as a private
helper function in an anonymous namespace in profile_manager.cc, so
we can't keep using the same approach.

Since we can't do that, we check the value of prefs::kProfileLastUsed
on startup and replace it with chrome::InitialProfile if we find that
its value is tor::kTorProfileDir, so there's no need to check things.

Additionally, we adapt a bit the browser tests to be able to check
for this migration, which happens very early on startup and therefore
we could not check it from the test once it's running.

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/c404e341e3693efe7c71c91da55de8b54d3bde66

commit c404e341e3693efe7c71c91da55de8b54d3bde66
Author: Alex Ilin <alexilin@chromium.org>
Date:   Tue Apr 6 12:02:53 2021 +0000

    [profiles] Clean up ProfileManager::GetLastUsedProfileName()

    "ProfileName" is ambiguous because it can refer to both
    - the profile directory within the user data directory, and
    - the visible profile name.

    This CL renames GetLastUsedProfileName() to
    GetLastUsedProfileBaseName(). In addition, this method was only used
    internally, and this CL moves it to anonymous namespace in the .cc file.

    Bug: 1195201