[Android] Implementation for SigninManagerImpl was changed
C-tor for SigninManagerImpl is private now and we can't extend it
without patching. Since we don't use it, here we just add our own mock
implementation of SigninManager interface.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/1bdcfc90ebdb8085b0ce348db5c5461472e22835

[Signin][Android] Move operations out of SigninManager ctor

This CL moves operations out of the SigninManager ctor to make the code
safer, otherwise we could get back to the object with it only partially
initialized.

Bug: 1152333