Make Windows silent installer truly silent (#8364)
Before this commit, there were two silent installers: a 1 MB online
"stub" installer and a standalone installer. Both opened a Brave window,
so were not truly silent. To fix this, the online installer was removed
entirely. And the standalone installer was updated to be built from a
new version of the Omaha update framework, which passes the flag
`--do-not-launch-chrome` to Brave's installer.