Drop patch to friend BrowserActionsContainer with BraveToolbarView
BrowserActionsContainer has been removed from upstream so this patch
won't apply. However, it seems that what the patch did (friending
that class with BraveToolbarView) is no longer needed because it
was added in commit 8013bbdc[1] to access the GetWidthForMaxWidth()
private method which hasn't been referenced from any Brave code for
a while now, and there doesn't seem to be any other need in the
BraveToolbarView class to access any of the other methods either.

Therefore we can safely delete this patch now.

[1] https://github.com/brave/brave-core/commit/8013bbdc

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/d35eeb9c748010219fd9ffc161996a67f4b53666

commit d35eeb9c748010219fd9ffc161996a67f4b53666
Author: Devlin Cronin <rdevlin.cronin@chromium.org>
Date:   Tue Apr 6 23:44:57 2021 +0000

    [Extensions UI] Fully Remove BrowserActionsContainer

    The BrowserActionsContainer was used in the legacy extensions toolbar
    implementation. It is no longer used or needed. Remove it entirely.

    Bug: 1165609