Cache mojo remotes to brave_shields::mojom::BraveShields when possible
Binding a new mojo remote each time a remote call is performed from the
BraveShieldsWebContentsObserver is slow and wasteful. Instead, we can
keep a map of lazily initialized mojo remotes that are created on
demmand the first time they are needed and reused the following times.

Note that we don't setup a disconnect handler in this case to prevent
a race condition that might happen if such handler was to be executed
after the RenderFrameHost was already deleted. Fortunately, this is
not needed anyway because the only situation in which a previously
bound remote will be disconnected would be precisely on deletion of
the RenderFrame in the renderer process, which will already get the
cached mojo removed from the map via the RenderFrameDeleted() callback.