[MacOS] Override helper-alerts identifier for signing.
Alerts helper has been added for signing upstream but Chrome doesn't
distribute it yet. Because it uses the same identifier as the current
alerts service, we get a signing error.

For now we'll use a different identifier for this helper and then
revert once Chrome starts distributing the helper.

Fixes brave/brave-browser#15731

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/56915522276f

commit 56915522276f86acc55f86b143b4d7b260d6bc36
Author: Richard Knoll <knollr@chromium.org>
Date:   Fri Mar 5 16:07:34 2021 +0000

    Introduce alert notification helper .app

    This adds a new helper .app on macOS to display alert notifications.
    This app is required to show alert style notifications as the main app
    can only show banner style ones and the XPC service can not use the new
    UNNotification APIs.

    Bug: 1127306