Make find_candidates() more robust
In the preprocess step, we used to delete elements with tags that
indicated the content is noise, such as footers, sidebars, and
ad-breaks. There have been a couple sites, such as the aclu, where the
main content is nested under one of these bad tags. So we re-run the
find_candidates() algorithm in 3 cases:
  1. We couldn't find any top candidate
  2. The top candidate has no content
  3. The top candidate's score is less than 5.

Resolves https://github.com/brave/brave-browser/issues/15887