Follow upstream TabContextMenuContents::GetAcceleratorForCommandId()
fix https://github.com/brave/brave-browser/issues/15894

Not 100% sure tab_ is the cause of this crash but it's the only difference with upstream.
We fetched widget via tab_ pointer but upstream fetches it via tabstrip_->GetWidget().
In this PR, BrowserView is used to get widget pointer because can't access
controller_->tabstrip_.