Add operational profiling collection for new federated learning component.
Operational profiles are anonymous, minimal representations of how the user engage with the browser over a collection period. The information from this one-off collection will guide the design of our federated learning infrastructure.

BraveFederatedLearningService acts as a controller for the federated learning tooling (for now only BraveOperationalProfiling). Operational profiles consist of a series of pings in 30m intervals to understand the potential participation of clients in a private federated learning protocol. To do so, BraveOperationalProfiling uses two timers: one firing at least once per collection epoch, and another timer firing a specified amount of time after the first has fired in order to simulate the time needed to compute an hypothetical model update by the client. A collection id, a user-generated random 128-bit number  with a fixed expiration, is attached to the log to allow the composition of the anonymous engagement data at the server.