speedreader: ui: Implement location bar bubbles
There are two bubble classes implemented in this PR:
`SpeedreaderBubbleSinglePage` and `SpeedreaderBubbleGlobal`. The former is
when Speedreader is not enabled. The UI prompts the user to enable
Speedreader. The latter is when Speedreader is enabled and shows the
current domain that is distilled, giving the user the option to
blacklist just that domain.

Both the classes share an interface in `SpeedreaderBubbleView`, which is
managed by the `SpeedreaderBubbleController` singleton.

To make the review process easier, the actual logic in both views is
stubbed out, and all the code is unreachable outside of tests. The
reason for this is because it will require changes to the Speedreader
service and the settings page, making this to hard to review on its own.

Resolves https://github.com/brave/brave-browser/issues/16162