Fix "declaration requires an exit-time destructor" compilation error
This happens due to adding these source files into //chrome/browser
due to stricter compilation checks apply to them now. Unfortunately,
we can't use base::NoDestructor<T> here since the offending variables
are meant to be set multiple times, so we need to go for the other
pattern commonly used in Chromium now, that is, to store a pointer
to the testing callback instead of the callback itself, and adapt
the signature and the callers (i.e. the test) for that.