Early return in BraveP3AService::OnHistogramChanged() if needed.
Early return in BraveP3AService::OnHistogramChanged() instead of
DCHECK-ing to prevent crashes that will happen for some histograms
that we found would be reported twice for the actual same event
as part of the same sequence, calling crashes on DCHECK-enabled builds.

So far, the events investigated were 3:

  Brave.NTP.SponsoredImagesEnabled
  Brave.SpeedReader.ToggleCount
  Brave.SpeedReader.Enabled

And in the three cases the actual same event (e.g. finishing the
initialization of the profile, restoring a previous session) ended
up trying to report the same histogram twice for the same event,
resulting on the histogram being properly sent on the first go
and then crashing on the second attempt (which would no longer be
necessary).

This patch simply replaces the DCHECK with an early return to make
sure that we don't crash on that second attempt when it happens,
and simply gracefully return instead.