[cr92 followup] Fixes blank settings page.
The page is blank because the /safetyCheck menu item cannot be found.
This happens because the menu item is no longer a direct child of the
menu, but instead is nested in a template.

Added code to search templates for menu items (perhpas there's a more
efficient way?).

Fixes brave/brave-browser#16708

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/49cea54a

commit 49cea54a094586e441513055fe0febedba4f377d
Author: dpapad <dpapad@chromium.org>
Date:   Fri May 14 08:53:26 2021 +0000

    Settings: Combine safetyCheck and privacy sections in re-design.

     - Hide the "Safety check" menu entry
     - Add a new |nest-under-section| attribute on <settings-section> and
       leverage it in main_page_behavior.js to show multiple
       <settings-section> instances at the same time.
     - Update tests.

    Bug: 1204457