Cleanup usage of base::Optional and base::nullopt #16589
These types are deprecated and we should be using absl::optional and
absl::nullopt now and avoiding the old ones in newly added code, so that
we don't have to keep changing that when rebasing the cr93 branch.

Also, we can now remove the temporary include rule in brave/DEPS to
allow //third_party/abseil-cpp/absl/types/optional.h, since such rule
is now available upstream.

Resolves https://github.com/brave/brave-browser/issues/16589