speedreader: Change how the icon behaves
After a meeting with the Speedreader team, Raf suggested we have too
much going on in the omnibox, and that we ought to kill the labels.
Instead of indicating page state via labels, we do so with different
icons. Icon vector change is triggered from the Update() method.

We get to delete lots of code now that we don't need to bother with
labels :)

Resolves https://github.com/brave/brave-browser/issues/9262