Fix Android Debug x86_64 build failures while linking
Android Debug build for x86_64 [1] is currently failing with these kind
of errors while linking:

  [51/503] SOLINK ./libmonochrome.cr.so
  FAILED: libmonochrome.cr.so libmonochrome.cr.so.TOC lib.unstripped/libmonochrome.cr.so
  python "../../build/toolchain/gcc_solink_wrapper.py"
    --readelf="../../third_party/android_ndk/toolchains/llvm/prebuilt/linux-x86_64/bin/x86_64-linux-android-readelf"
    --nm="../../third_party/android_ndk/toolchains/llvm/prebuilt/linux-x86_64/bin/x86_64-linux-android-nm"
    --strip=../.>
  ld.lld: error: undefined symbol: ChromeContentClient::ChromeContentClient()

The problem is a missing //brave/common->//chrome/common dependency,
but we can't simply add it or we'll get a dependency cycle.

Fortunately, the only cases that include files from //brave/common are
chromium_src sources overriding files not included in //chrome/common:

  * chrome/common/channel_info_posix.cc, from //chrome/common:channel_info
  * chrome/common/chrome_paths_linux.cc, from //chrome/common:constants

The first case already depends on //brave/common:channel_info, so all
we have to do is to add that dependency in //chrome/common:constants
as well so we can drop the //chrome/common->//brave/common dependency,
and fix both the Android build and this potential deps cycle in one go.

[1] npm run build -- Debug --target_os=android --target_arch=x64

Resolves: https://github.com/brave/brave-browser/issues/11909