[bat-native-ledger] Cleanup usage of base::Optional
This type is deprecated and we should be using absl::optional now and
avoiding the old one in newly added code, so that we don't have to keep
changing that when rebasing the cr93 branch.

This CL removes the few old types remaining in master after the main
cleanup done at https://github.com/brave/brave-core/commit/1c989344.

Resolves https://github.com/brave/brave-browser/issues/16589