Fix EthNonceTrackerUnitTest.* unit tests on Linux Component builds
Despite of these tests passing on CI without any issues, they are
consistenly failing at least on Linux Component builds and the issue
seems to be related to the ethereum blockchain network used in the
tests: if Mainnet is used and a full valid URL is specified by
including the PATH part in the URL (e.g. [1]), the tests fail.

Removing the PATH part from the URL and simply using something such
as https://mainnet-infura.brave.com/ would get the tests passing,
but probably the right fix is something like what we did in PR 9388
and switch to using the "localhost network" for tests instead, which
is what this CL implements.

[1] https://mainnet-infura.brave.com/f7106c838853428280fa0c585acc9485
[2] https://github.com/brave/brave-core/pull/9388

Resolves https://github.com/brave/brave-browser/issues/16992