Fix npm run tslint errors due to no-floating-promises being violated
With the upgrade of the @types/chrome node module to 0.0.147, we'd also
be pulling a change in the linter that will now recognize functions
returning Promises that were not identified before, and which we were
not handling properly before (e.g. chrome.tabs.executeScript()),
resulting in lint errors like this:

  ERROR: src/brave/components/brave_extension/extension/brave_extension/background/api/browserActionAPI.ts:33:3 - Promises must be handled appropriately
  ERROR: src/brave/components/brave_extension/extension/brave_extension/background/api/browserActionAPI.ts:43:3 - Promises must be handled appropriately
  ERROR: src/brave/components/brave_extension/extension/brave_extension/background/api/browserActionAPI.ts:52:5 - Promises must be handled appropriately

  [...]

  ERROR: src/brave/components/brave_wallet_ui/panel/async/wallet_panel_async_handler.ts:80:3 - Promises must be handled appropriately
  ERROR: src/brave/components/brave_wallet_ui/panel/async/wallet_panel_async_handler.ts:84:3 - Promises must be handled appropriately
  ERROR: src/brave/components/brave_webtorrent/extension/components/torrentViewerHeader.tsx:70:9 - Promises must be handled appropriately

This patch silences the linter by making sure that the returned
Promises are accounted for using the .catch() method to log an
error in case something went wrong.