Fixes text in Additional Settings button being cut off.
Fixes brave/brave-browser#17109

The button host style is set to have a 100px border radius that
cuts off the text when the button transitions to multiline in
certain languages (DE, FR).

The radius is set in ui/webui/resources/overrides/cr_button.js
Added a style override for the #advancedButton in our menu override.