speedreader: Point to speedreader in settings
We have two options here: (1) Point to
brave://settings?search=speedreader, which will highlight the
Speedreader setting, or (2) or provide a pointer to
brave://settings/appearance. I went with the former because it makes the
setting a bit easier to identify. There isn't a way to link to a
specific setting.

Resolves https://github.com/brave/brave-browser/issues/17216