[iOS] Don't rename mojo-generated Obj-C protocols in Swift
In Swift, you cannot have a nested protocol, but there is a bug with Xcode that lets you still compile renaming a protocol to be nested via `NS_SWIFT_NAME`. In Swift this protocol cannot be conformed to if the protocol itself uses any of the other nested types since Swift expects all protocols to be top-level and doesn't use the proper renamed type within the method definitions.