Reverts usage of JS APIs that return promises.
These were introduced by https://github.com/brave/brave-core/pull/9545,
however, according to Chrome documentation, e.g.:
* https://developer.chrome.com/docs/extensions/reference/tabs/
* https://developer.chrome.com/docs/extensions/reference/browserAction/
the API versions that return Promise are only available to extensions
that use Manifest V3. We are using Manifest V2 and should instead be
using APIs that take a callback.

Replaced all usages of these APIs that expected a Promise returned with
the versions that take a callback.

Fixes brave/brave-browser#17654

Additionally, added some return statements inside .then() in
shieldsPanelReducer.ts to get rid of the Bluebird warnings that "a
promise was created in a handler but was not returned from it".