Work around UI deps being weird for wallet page
The wallet page needs to find the right browser so it can show a popup
panel when the page requests it.  This should normally be possible but
our deps are weird in the UI level right now.

I discussed this with BrianJ and Mario and they were OK with this approach of collapsing this BUILD.gn down.