speedreader mojo structure concept
We create two mojo types in this PR: SpeedreaderUI and
SpeedreaderResult.

SpeedreaderResult is for communicating the distillation result from the
browser (speedreader_tab_helper) to the renderer_observer. If the result
is distilled, then the renderer initializes its JavaScript handler.

SpeedreaderUI is how the renderer communicates with the browser process
for live UI update. It will request all the current preferences upfront
on load, then register itself as an observer to be notified if the
preferences change.