Initial commit for SKU SDK component
Goal is to expose a few SKU SDK methods via render frame observer injection:
- `window.brave.skus.refresh_order`
- `window.brave.skus.fetch_order_credentials`
- `window.brave.skus.prepare_credentials_presentation`
- `window.brave.skus.credential_summary`

If these methods are detected by the JS SDK on the account.brave.com website, they can be
called in lieu of the methods in the JS SDK - providing a way for the credential to be
safetly be intercepted by the browser.

The first use-case for using a credential is with VPN. The VPN credential is read and
then passed through to the vendor.

`brave_content_browser_client` connects the renderer side
(`BraveSkusRenderFrameObserver` and `BraveSkusJSHandler`) to the
implementation (SkusSdkImpl) via mojom bindings (see
`RegisterBrowserInterfaceBindersForFrame`).

The implementation in this commit exposes the `refresh_order` method and
returns a dummy JSON response.