BraveSessionCache: Update farbling browser tests
Update the expected values in the affected brave_browser_tests
for the values from the new generator. Since these rely on
deterministic generation, but values are stable from run to
run, but depend on the prng used and its seeding.

The new values we copied directly from the failing test output.