Revert "Revert "[MacOS][Signing] Revert overwriting Alerts helper app id.""
This reverts commit a0eb74ac78e1f97fc401ccd01c8d2f93d5496d0b.

We don't need to modify Alert Helper identifier after all. The reason we
were getting the signing error after upstream switched to using the
helper instead of the XPC service for the notification was because we
were accidentally clobbering the Alert Helper plist file. This was fixed
in https://github.com/brave/brave-core/pull/10172 and so now we can
revert the revert of the revert and sign Alert Helper normally.

Fixes brave/brave-browser#18282