SearchEngineTracker: Re-enable DefaultSearchEngineP3A test.
This was disabled in 27bf9e3b0b97 for intermittent failures.
Add a call to wait for TemplateURLServiceFactory to finish
loading, and re-enable to see if that stabilizes the browser
test.

See https://github.com/brave/brave-browser/issues/13057 for tracking.

This duplicates VerifyTemplateURLServiceLoad() from the
search_engine_provider browser test, but it's a small piece
of code and having two file-local implementations is simple
enough.